services/zanasonic-etl/pyproject.toml
[tool.poetry]
name = "zanasonic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.11"
scipy = "^1.8.1"
pandas = "^1.4.2"
loguru = "^0.6.0"
scikit-learn = "^1.1.1"
requests = "^2.28.0"
pyarrow = "^8.0.0"
pyjanitor = "^0.23.1"
pydantic = "^1.9.1"
strictyaml = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
isort = "^5.10.1"
black = "^22.3.0"
pylint = "^2.14.2"
jupyterlab = "^3.4.3"
ipykernel = "^6.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
